Move PhysicalQubit trait.
